+++ libcnb/src/target.rs
@@ -1,4 +1,4 @@
-#[derive(Debug)]
+#[derive(Clone, Debug)]
 pub struct Target {
     /// The name of the target operating system.
     ///