+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `libcnb-data':
+- `libcnb-data`:
   - Reintroduced support for deserializing `[[stacks]]` in `buildpack.toml`.
     ([#789](https://github.com/heroku/libcnb.rs/pull/789))
 - `libherokubuildpack`: