+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `libcnb`:
-  - `Target` now implements `Debug`. ([#821](https://github.com/heroku/libcnb.rs/pull/821))
+  - `Target` now implements `Clone` and `Debug`. ([#821](https://github.com/heroku/libcnb.rs/pull/821))
 
 ### Changed