+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.21.0] - 2024-04-30
+
 ### Added
 
 - `libcnb`:
@@ -289,7 +292,8 @@ version number. See the changelog below for other changes.
 
 - Remove support for legacy BOM. Remove `Launch::bom`, `Build::bom`, `bom::Bom`, `bom::Entry`. ([#489](https://github.com/heroku/libcnb.rs/pull/489))
 
-[unreleased]: https://github.com/heroku/libcnb.rs/compare/v0.20.0...HEAD
+[unreleased]: https://github.com/heroku/libcnb.rs/compare/v0.21.0...HEAD
+[0.21.0]: https://github.com/heroku/libcnb.rs/compare/v0.20.0...v0.21.0
 [0.20.0]: https://github.com/heroku/libcnb.rs/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/heroku/libcnb.rs/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/heroku/libcnb.rs/compare/v0.17.0...v0.18.0